move out file data access operation from database.js
