+++ templates/upload.html
@@ -11,7 +11,7 @@
     <tr>
       <td>Data Type</td>
       <td>
-        <select id="type" title="select upload file type">
+        <select id="type" class="input-selectbox" title="select upload file type">
           <option value="network">Network</option>
           <option value="binding">Binding Track</option>
           <option value="bed">Bed Track</option>
